+++ src/extension.ts
@@ -33,7 +33,10 @@ export function activate(context: vscode.ExtensionContext) {
   /// Register the commands
   commands.registerCommand(Commands.resetColors, resetColorsHandler);
 
-  commands.registerCommand(Commands.saveColorToFavorites, saveColorToFavoritesHandler);
+  commands.registerCommand(
+    Commands.saveColorToFavorites,
+    saveColorToFavoritesHandler
+  );
 
   commands.registerCommand(Commands.enterColor, enterColorHandler);
 
@@ -67,11 +70,11 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.workspace.onDidChangeConfiguration(async e => {
       if (checkIfPeacockSettingsChanged(e) && state.recentColor) {
-        console.log(
-          `Configuration changed. Changing the color to most recently selected color: ${
-            state.recentColor
-          }`
-        );
+        // console.log(
+        //   `Configuration changed. Changing the color to most recently selected color: ${
+        //     state.recentColor
+        //   }`
+        // );
         await changeColor(state.recentColor);
       }
     })